Move GUDateTimeFormat to common model
