+++ common/app/views/fragments/meta/dateline.scala.html
@@ -1,5 +1,6 @@
 @import org.joda.time.DateTime
-@import views.support.{GUDateTimeFormat, Format}
+@import model.GUDateTimeFormat
+@import views.support.{Format}
 
 @(webPublicationDate: DateTime, lastModified: DateTime, hasBeenModified: Boolean, firstPublicationDate: Option[DateTime], isLiveBlog: Boolean = false, isLive: Boolean = false, isMinute: Boolean = false)(implicit request: RequestHeader)
